requirements.txt
openai>=1.0.0
pandas>=1.5.0
plotly>=5.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
python-docx>=0.8.11
weasyprint>=60.0
matplotlib>=3.5.0
squarify>=0.4.3